+++ appinfo/application.php
@@ -1,4 +1,14 @@
 <?php
+/**
+ * ownCloud - files_mv
+ *
+ * This file is licensed under the Affero General Public License version 3 or
+ * later. See the COPYING file.
+ *
+ * @author eotryx <mhfiedler@gmx.de>
+ * @copyright eotryx 2015
+ */
+
 
 namespace OCA\Files_mv\Appinfo;