+++ CHANGELOG.md
@@ -1,2 +1,25 @@
-owncloud-filesmv (0.0.1)
-* First release
+CHANGELOG
+=========
+
+0.8.0
+-----
+
+ * Licence change to AGPL
+ * Unit tests for MoveController (CompleteController test is shit)
+ * translation of various languages
+
+0.7.3a
+------
+
+ * fix xml error in appinfo/info.xml
+
+0.7.3
+-----
+
+ * initial OC8 support
+
+ missing features:
+  
+  * translation
+  * unit tests
+